+++ back-end/config/post.go
@@ -1,0 +1,7 @@
+package config
+
+func (p Post) NewPost() {
+	// Commandes pour enregistrer dans la db
+	print("test")
+	return
+}
