+++ back-end/config/struct_.go
@@ -14,3 +14,13 @@ type User struct {
 	Cookie          string
 	CookieLifeTime  string
 }
+
+type Post struct {
+	Id           string
+	UserId       string
+	Content      string
+	Visibility   string
+	AllowedUsers []string
+	Image        string
+}
+
